requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==4.2.2
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.2
attrs==25.4.0
audio-recorder-streamlit==0.0.10
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==41.0.7
distro==1.9.0
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
ecdsa==0.19.1
entrypoints==0.4
fastapi==0.109.0
filelock==3.24.2
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.62.3
gTTS==2.5.4
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
huggingface_hub==0.36.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.9
langchain-core==1.2.9
langchain-openai==1.1.8
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.7.1
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
mysql-connector-python==8.0.33
networkx==3.6.1
nltk==3.9.2
numpy==1.26.4
ollama==0.6.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.1.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==3.20.3
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.23
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pydeck==0.9.1
PyJWT==2.10.1
pymongo==4.16.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.13.1
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==2.2.2
sentencepiece==0.2.1
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.3
SpeechRecognition==3.14.5
SQLAlchemy==2.0.23
sqlparse==0.5.3
srt==3.5.3
starlette==0.35.1
streamlit==1.53.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
toolz==1.1.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.6
twilio==9.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.27.0
vosk==0.3.45
watchdog==6.0.0
watchfiles==1.1.1
webrtcvad-wheels==2.0.14
websockets==12.0
Werkzeug==3.1.5
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
python-decouple
python-jose
httpx==0.27.2
openai
